% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbReconnect.R
\name{class_is_oracle}
\alias{class_is_oracle}
\title{Class Is Oracle}
\usage{
class_is_oracle(x)
}
\arguments{
\item{x}{an object or string (preferabbly a character string) to test to see if it's an Oracle connection}
}
\description{
Class Is Oracle
}
