% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_date.R
\name{oracle_date}
\alias{oracle_date}
\title{Convert R Dates to Oracle Date}
\usage{
oracle_date(a_date)
}
\arguments{
\item{a_date}{a date or POSIXct object (everything passed gets run through as.Date() before being passed to Oracle)}
}
\value{
a to_date() function which should work in Oracle
}
\description{
pass an R date object to Oracle as a date
}
